scrapp/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Login==0.5.0
Flask-WTF==0.14.3
Flask-SocketIO==5.0.1
Flask-Caching==1.9.0
Flask-Migrate==2.5.3
Flask-Assets==2.0
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
cssmin==0.2.0
email-validator==1.1.2
PyJWT==2.0.0
gevent==20.9.0
gevent-websocket==0.10.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
requests==2.24.0
boto3==1.16.10
redis==3.5.3
python-json-logger==2.0.1
structlog==21.1.0
colorama==0.4.4
elasticsearch==7.11.0